feat: compatible with X.Org Server 21.1
This commit is authored with assistance from AI/LLM:

- Model: DeepSeek V4 Flash
- Platform: DeepSeek Platform
- Agent platform: Kilo Code
- Prompt:
	- 1.这是一个GPU驱动的源码，且参考该目录与根目录源码的区别，使其可以同时支持M2012和M2015这两款GPU IP
	- 2.现在请尝试修复并适配Xorg 21.1，使其可以在较新的发行版上成功编译并使用
	- 3.该GPU IP仅出现在32位处理器上，且仅支持32位操作系统，当前修改是否有影响？
	- 4.不需要兼容老Xorg版本，请为我更新make clean命令清理多余文件
	- 5.@log/Xorg.0.log 顺利安装驱动后无法启动xorg server，这是返回的日志，请检查原因
	- 6.@misc/xorg.conf @misc/RDCVBIOS.ROM 其中xorg.conf来自于GPU厂商驱动，分析该文件是否有用

find more information in docs/llm-session.md.

+++ src/BiosDef.h
@@ -26,11 +26,11 @@ typedef unsigned char   CBStatus;
 
 typedef unsigned char   UCHAR;
 typedef unsigned short  WORD;
-typedef unsigned long   DWORD;
+typedef unsigned int    DWORD;
 typedef char*           PCHAR;
 typedef unsigned short  USHORT;
 typedef long            LONG;
-typedef unsigned long   ULONG;
+typedef unsigned int    ULONG;
 typedef unsigned long long   uint64;
